+++ scripts/build-wasm.sh
@@ -5,8 +5,7 @@
 
 set -e
 
-# also change in install-build-deps
-RUST_VER=nightly-2020-10-25
+source ../scripts/rust-version.sh
 
 WSM=bytes/wasm_bg.wasm
 OPT=bytes/wasm_opt.wasm
@@ -15,17 +14,18 @@ ASM=wasm-crypto-asmjs/build/cjs/data.js
 # build new via wasm-pack
 echo "*** Building Rust sources"
 cd wasm-crypto
-rustup run $RUST_VER xargo build --target wasm32-unknown-unknown --release
-# rustup run $RUST_VER cargo build --target wasm32-unknown-unknown --release -Z build-std=std,panic_abort
+if [ "$RUST_VER" == "stable" ]; then
+  RUSTC_BOOTSTRAP=1 cargo build --target wasm32-unknown-unknown --release -Z build-std=std,panic_abort
+else
+  rustup run $RUST_VER xargo build --target wasm32-unknown-unknown --release
+fi
 cd ..
 
 echo "*** Converting to WASM"
 ../bindgen/wasm-bindgen wasm-crypto/target/wasm32-unknown-unknown/release/wasm.wasm --out-dir bytes --target web
-# wasm-pack build --release --scope polkadot --target web
 
 # optimise
 echo "*** Optimising WASM output"
-# ../../wabt/bin/wasm-strip $WSM
 ../binaryen/bin/wasm-opt $WSM -Os -o $OPT
 
 # convert wasm to base64 structure